requirements.txt
pyserial
pandas
matplotlib
pyxdf
mne
scipy
pylsl